requirements.txt
Django==2.1.2
Pillow==2.6.1
Unipath==1.1
aiodataloader==0.1.2
coverage==4.5.1
django-appconf==1.0.2
django-cors-headers==2.4.0
django-discover-runner==1.0
django-imagekit==4.0.2
django-redis==4.9.0
django-cachalot==2.1.0
django-cursor-pagination==0.1.3
django-tastypie==0.14.2
djangorestframework==3.9.0
drf-ujson==1.2.0
graphene==2.1.3
graphene-django==2.2.0
gunicorn==19.9.0
mimeparse==0.1.3
pilkit==2.0
psycopg2==2.7.5
python-dateutil==2.7.3
python-mimeparse==1.6.0
simplejson==3.16.0
six==1.11.0
